SURFER Service, Characteristics, and Response Handlers Setup
In this commit the service and characteristics for the SURFER were given a clear outline and data path, as well as preparing for future handlers. The inventory and initialize buttons should be able to send data to the SURFER. If an RFID tag is detected by the SURFER, the raw data packets will be displayed using toast messages.

Possible Risks/Issues to be Verified:

The Listening Characteristics in updateCharacteristics need to have their notifications enabled.

Characteristics in the original list aren't automatically updated by the onCharacteristic functions.

Data to be sent is being invoked improperly, and causing the MCU to not respond appropriately if at all.