Application State, Send Blank EPC, and Characteristic Notifications
In this commit the Application State Machine was completely fleshed out aside from the Recovering Waveform Memory FIFO. In addition to this, the Write Target EPC button was partially added as a btnSendBlankEPC. This allows for us to tell the reader to remove all filters for Tag searches before starting an inventory.

The Characteristics notifications were enabled to ensure receival of data, as it was determined they're not enabled by default. This has fixed a previously identified risk in the "SURFER Service, Characteristics, and Response Handlers Setup" commit.

Possible Risks/Issues:

The characteristic notifications aren't functioning properly.

The blank EPC is formatted incorrectly, causing the Reader to not change its filters.

The inventory is currently untested, and the ability for the reader to detect RFID Tags is unkown.

The receival of the tag responses may not be displayed or recognized properly using the Toast messages.